requirements.txt
# TEMU 智能出图系统 - Python 依赖
# 核心作者: 企鹅

# Web 框架
streamlit>=1.29.0

# 图片处理
Pillow>=10.0.0

# Google AI
google-genai>=0.3.0

# 环境变量
python-dotenv>=1.0.0
